refactor: move page content to separate components
